# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface NamedItem {
  id: number;
  name: string;
}

export interface Gear {
  id: number;
  name: string;
  level: number;
}

export interface PlayerBrawler {
  id: number;
  name: string;
  power: number;
  rank: number;
  trophies: number;
  highestTrophies: number;
  gadgets: NamedItem[];
  starPowers: NamedItem[];
  gears: Gear[];
  hypercharges: NamedItem[];
}

export interface Player {
  tag: string;
  name: string;
  /** @nullable */
  nameColor: string | null;
  iconId: number;
  trophies: number;
  highestTrophies: number;
  expLevel: number;
  expPoints: number;
  soloVictories: number;
  duoVictories: number;
  trioVictories: number;
  /** @nullable */
  clubTag: string | null;
  /** @nullable */
  clubName: string | null;
  /** @nullable */
  clubRole: string | null;
  brawlers: PlayerBrawler[];
}

export interface BattlePlayer {
  tag: string;
  name: string;
  brawlerId: number;
  brawlerName: string;
  brawlerPower: number;
  brawlerTrophies: number;
}

export interface BattleLogEntry {
  battleTime: string;
  mode: string;
  /** @nullable */
  map: string | null;
  type: string;
  /** @nullable */
  result: string | null;
  /** @nullable */
  duration: number | null;
  /** @nullable */
  trophyChange: number | null;
  /** @nullable */
  starPlayerTag: string | null;
  teams: BattlePlayer[][];
}

export interface BattleLog {
  items: BattleLogEntry[];
}

export interface TodayStats {
  date: string;
  gamesPlayed: number;
  wins: number;
  losses: number;
  draws: number;
  timePlayedMinutes: number;
  trophyChange: number;
}