# lib/api-zod/src/generated/types/battleLog.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BattleLogEntry } from './battleLogEntry';

export interface BattleLog {
  items: BattleLogEntry[];
}