# lib/api-zod/src/generated/types/battleLogEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BattlePlayer } from './battlePlayer';

export interface BattleLogEntry {
  battleTime: Date;
  mode: string;
  /** @nullable */
  map: string | null;
  type: string;
  /** @nullable */
  result: string | null;
  /** @nullable */
  duration: number | null;
  /** @nullable */
  trophyChange: number | null;
  /** @nullable */
  starPlayerTag: string | null;
  teams: BattlePlayer[][];
}