# lib/api-zod/src/generated/types/battlePlayer.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BattlePlayer {
  tag: string;
  name: string;
  brawlerId: number;
  brawlerName: string;
  brawlerPower: number;
  brawlerTrophies: number;
}