# lib/api-zod/src/generated/types/gear.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Gear {
  id: number;
  name: string;
  level: number;
}