# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerBrawler } from './playerBrawler';

export interface Player {
  tag: string;
  name: string;
  /** @nullable */
  nameColor: string | null;
  iconId: number;
  trophies: number;
  highestTrophies: number;
  expLevel: number;
  expPoints: number;
  soloVictories: number;
  duoVictories: number;
  trioVictories: number;
  /** @nullable */
  clubTag: string | null;
  /** @nullable */
  clubName: string | null;
  /** @nullable */
  clubRole: string | null;
  brawlers: PlayerBrawler[];
}