# lib/api-zod/src/generated/types/playerBrawler.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Gear } from './gear';
import type { NamedItem } from './namedItem';

export interface PlayerBrawler {
  id: number;
  name: string;
  power: number;
  rank: number;
  trophies: number;
  highestTrophies: number;
  gadgets: NamedItem[];
  starPowers: NamedItem[];
  gears: Gear[];
  hypercharges: NamedItem[];
}