# lib/api-zod/src/generated/types/todayStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TodayStats {
  date: string;
  gamesPlayed: number;
  wins: number;
  losses: number;
  draws: number;
  timePlayedMinutes: number;
  trophyChange: number;
}